fix: update catalystcommunity dependency version (#13)
there isn't a release that contains the new go.mod file changes from the
org migration. this change should trigger a release that should fix that
problem.

additionally, this change updates the only catalystcommunity dependency
of this go project to prevent problems when importing the package from
the new org url.